fixed off-by-one issue in repeat label
Previously, patterns like '{2}' would be labelled 'once',
and '{0}' could end up as '-1 Times'.
The repeat-label logic was adjusted, along with special
handling for the non-sensical '{0}' case.
Also I am not quite sure how it should be visualized, right
now 'a{0}' looks nearly similar to 'a?', even though it is
quite different effectively.

Fixes #19